+++ Sources/SwiftUIKit/_Deprecated/FontListPicker.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "SystemFontListPicker")
+public typealias FontListPicker = SystemFontListPicker